update mutt change_box.sh

+++ mutt/change_box.sh
@@ -2,9 +2,12 @@
 
 selected="$HOME/.mutt/mail/selected"
 
-# TODO - make this more robust
+# read box
 box="$HOME/.mutt/mail/$(find $HOME/.mutt/mail -mindepth 2 -maxdepth 2 | cut -d '/' -f 6- | grep -v '^\.' | fzf)"
 
-rm -f $selected
+if [[ "$box" == "" ]]; then
+    exit
+fi
 
+rm -f $selected
 ln -s $box $selected
